Create userDictionary in ChatListController
The userDictionary allows for fast and easy lookup of all of the currentUser's chat conversations. The code sample obtained from C. Talha would save the actual senderName into Firestore. In our app, it's possible for a user to change their first name or last name, therefore, I needed a way at run time to determine the names of all the users the currentUser is chatting with. The userDictionary is constructed in two phases. First, as soon as currentUser is didSet in ChatListController, we add the currentUser's UID (key) and firstName (value). Second, during the fetchChatsOfCurrentUser in ChatListController, we append on all the other users who are chatting with the currentUser.