Fix bug that repeatedly appended duplicate chats
My understanding of Firestore Snapshot Listeners was wrong. If an add is detected in the difference, it doesn't necessarily mean that there's a brand new document. The Listener will always return all matching documents - including older documents - every time it is invoked, which is usually done whenever viewDidAppear happens. I validated from another developer's blog that your source of truth, the array for your TableViewController, must be emptied first before you add the diffs. The value-add for Listeners is therefore only when 2 conditions are met: 1) the user is looking at - and remains at - the VC which launched your Listener and 2) Firestore data has changed.

Another learning point: listeners substitute for a getDocuments call. There's no need for both assuming the queries for both are identical.